06/28 Mp3교체, 사운드 플레이 방식 코드 수정

+++ Assets/Scripts/Battle/Effect/EffectObject.cs
@@ -63,7 +63,7 @@ namespace Battle.Effect
 			//        }
 			//    }
 			//}
-			Sound.PlayEff((int)EffSoundType.Attack);
+			Sound.PlayEff(EffSoundType.Attack);
 		}
 		private void Update()
 		{
